utils: Fix sign compare warning
The result of snprintf() is int and it is compared against an size_t buffer
length. Add the missing cast. It is safe to assume that ret is positive at
this point.

+++ src/utils.h
@@ -225,7 +225,7 @@ static inline int snprintf_err_handling(char **buffer, size_t *len, int ret)
 		return -EINVAL;
 
 	/* Buffer too small? */
-	if (ret >= *len)
+	if ((size_t)ret >= *len)
 		return -EINVAL;
 
 	/* All good */
